bb.fetch2.{git, hg}: force tarball generation when repo is re-fetched
The fetcher expects to see a tarball at the end of fetching
if one is found under source mirrors. With the current
logic, if the fetched scm tarball does not contain the
required hashes the repo is re-fetched and the downloaded
tarball is deleted. This causes the fetcher to fail
as it looks for the tarball and only passes in such cases
when the BB_GENERATE_MIRROR_TARBALLS is set.
We now force generation of tarball irrespective of
BB_GENERATE_MIRROR_TARBALLS in cases where the
repo is re-fetched.

Signed-off-by: Awais Belal <awais_belal@mentor.com>